+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 dev:
   - fix incorrect error when "deposit veirfy" is not given a withdrawal address
+  - allow truncated mnemonics (first four characters of each word)
 
 1.31.0:
   - initial support for deneb